Move OpenPROnGitHub from View to ViewModel
